Refactored how parser adds a widget mark while parsing within yaml-tangram.js Before, functionality for adding widget markers was within widgets folder in widget.js. Part of that functionality is now stored within codemirror folder. The other part is stored within widget-manager.js
